tests/wpt/tests/html/canvas/element/text/2d.text.direction.invalid.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.text.direction.invalid</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.text.direction.invalid</h1>
<p class="desc"></p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("");
_addTest(function(canvas, ctx) {

  ctx.direction = 'ltr';
  ctx.direction = 'rtl ';
  _assertSame(ctx.direction, 'ltr', "ctx.direction", "'ltr'");

  ctx.direction = 'rtl';
  ctx.direction = 'LTR';
  _assertSame(ctx.direction, 'rtl', "ctx.direction", "'rtl'");

  ctx.direction = 'ltr';
  ctx.direction = 'rtl\0';
  _assertSame(ctx.direction, 'ltr', "ctx.direction", "'ltr'");

  ctx.direction = 'ltr';
  ctx.direction = 'bogus';
  _assertSame(ctx.direction, 'ltr', "ctx.direction", "'ltr'");

  ctx.direction = 'ltr';
  ctx.direction = 'inheri';
  _assertSame(ctx.direction, 'ltr', "ctx.direction", "'ltr'");

});
</script>
